+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - **README** Updated README
 - **Package Handler** Replace difflib get_close_matches for tag comparison with rapidfuzz process for better results
 - **General** Refactoring, improve GitLab API usage
+- **GitLab API** Refactoring, implement retry mechanic, add a function that extracts certain information from the upstream url
 
 ### Bug fixes